+++ uploader/test/test_file_upload.py
@@ -2,51 +2,19 @@ import logging
 import tempfile
 import zipfile
 
-from django.utils import timezone
 from openpyxl.workbook import Workbook
 
-from core.models import Iso639LanguageCode
-from institution.models import CofkUnionInstitution
-from location.models import CofkUnionLocation
-from person.models import  CofkUnionPerson
 from uploader.constants import MANDATORY_SHEETS
-from uploader.models import CofkCollectUpload, CofkCollectStatus, CofkCollectWork, CofkCollectAuthorOfWork, \
+from uploader.models import CofkCollectWork, CofkCollectAuthorOfWork, \
     CofkCollectAddresseeOfWork, CofkCollectOriginOfWork, CofkCollectDestinationOfWork, CofkCollectManifestation
 from uploader.spreadsheet import CofkUploadExcelFile
-from uploader.test.test_utils import UploaderTestCase
+from uploader.test.test_utils import UploadIncludedTestCase
 from uploader.validation import CofkExcelFileError
 
 log = logging.getLogger(__name__)
 
 
-class TestFileUpload(UploaderTestCase):
-    def setUp(self) -> None:
-        super().setUp()
-        CofkCollectStatus.objects.create(status_id=1,
-                                         status_desc='Awaiting review')
-
-        CofkUnionLocation(pk=782).save()
-
-        for lang in ['eng', 'fra']:
-            Iso639LanguageCode(code_639_3=lang).save()
-
-        CofkUnionInstitution(institution_id=1,
-                             institution_name='Bodleian',
-                             institution_city='Oxford').save()
-
-        for person in [{'person_id': 'a', 'iperson_id': 15257, 'foaf_name': 'Newton'},
-                       {'person_id': 'b', 'iperson_id': 885, 'foaf_name': 'Baskerville'},
-                       {'person_id': 'c', 'iperson_id': 22859, 'foaf_name': 'Wren'}]:
-            CofkUnionPerson(**person).save()
-
-        CofkUnionLocation(location_id=400285).save()
-
-        self.new_upload = CofkCollectUpload()
-        self.new_upload.upload_status_id = 1
-        self.new_upload.uploader_email = 'test@user.com'
-        self.new_upload.upload_timestamp = timezone.now()
-        self.new_upload.save()
-        
+class TestFileUpload(UploadIncludedTestCase):
     def test_create_upload(self):
         self.assertEqual(self.new_upload.upload_status.status_desc, 'Awaiting review')
         self.assertEqual(self.new_upload.total_works, 0)
